--- client-compose.yml
@@ -1,7 +1,0 @@
-version: '3.3'
-
-services:
-  aigym:
-    environment:
-      FILE_PATH: ${FILE_PATH}
-    build: ./runtime-enviroment
