+++ clients-enviroment/requirements.txt
@@ -1,0 +1,1 @@
+google-cloud-storage
