+++ clients-enviroment/src/cloud_storage_controller.py
@@ -20,9 +20,9 @@ class CloudStorage:
 
 
 if __name__ == "__main__":
-    path_to_zipfile = "../src/files.zip"
+    path_to_zipfile = "../project/src/files.zip"
     storage = CloudStorage()
     storage.import_file(os.environ.get('FILE_PATH'), path_to_zipfile)
     with ZipFile(path_to_zipfile, 'r') as zip_ref:
-        zip_ref.extractall("../src/")
+        zip_ref.extractall("../project/src/")
