+++ clients-enviroment/src/start.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cd src
+python cloud_storage_controller.py
