+++ docker-compose.yml
@@ -1,0 +1,20 @@
+version: '3.3'
+
+services:
+  aigym:
+    build: ./runtime-enviroment
+    depends_on:
+      cloud-container:
+        condition: service_completed_successfully
+    volumes:
+      - project-volume:/app/project/src
+
+  cloud-container:
+    build: ./clients-enviroment
+    environment:
+      FILE_PATH: ${FILE_PATH}
+    volumes:
+      - project-volume:/app/project/src
+
+volumes:
+  project-volume:
