+++ runtime-enviroment/src/start.sh
@@ -1,11 +1,6 @@
 #!/bin/bash
-cd storage
-python cloud_storage_controller.py
-
-cd ..
+cd project/src
 pipreqs .
 pip install -r requirements.txt
-
-cd src
 python -u code.py
 # stdbuf -oL python code.py > log
